pyproject.toml
[tool.poetry]
name = "client-authentication"
version = "0.1.0"
description = "Client authentication system for QSAAS"
authors = ["Ahmad Khaled <36166958+ahmadkkhaled@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
django-extensions = "^3.1.3"
djangorestframework = "^3.12.4"
django-environ = "^0.4.5"
psycopg2 = "^2.9.1"
djangorestframework-simplejwt = "^4.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
